requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.0
colorama==0.4.6
fastapi==0.115.12
h11==0.16.0
idna==3.10
pydantic==2.11.4
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
# Database
sqlalchemy==2.0.31
alembic==1.13.1
psycopg2-binary==2.9.9
# Storage
boto3==1.34.86
python-dotenv==1.0.1
aiofiles==23.2.1
